separate CopyContextMenuItem from RevisionGrid
